refactor(elevation-service): update missing names from elevation to elevation service
